Upgrade to Scala 3

+++ src/main/scala/utils/PerfUtils.scala
@@ -9,7 +9,7 @@ object PerfUtils {
   def time[A](id: String)(block: => A): A = {
     val startTime = System.nanoTime
     val result = block
-    val elapsedMicro = Math.round((System.nanoTime() - startTime) / 1000)
+    val elapsedMicro = Math.round((System.nanoTime() - startTime) / 1000.toDouble)
     val elapsedStr = if (elapsedMicro > 5000) s"${elapsedMicro / 1000}ms" else s"${elapsedMicro}us"
     println(s"$id took $elapsedStr")
     result
